requirements.txt
# Core dependencies
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.25.2

# Authentication and security
flask-jwt-extended==4.5.3
bcrypt==4.0.1

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7  # For PostgreSQL
pymysql==1.1.0  # For MySQL (optional)

# API and data processing
finnhub-python==2.4.18
yfinance==0.2.31
matplotlib==3.8.0
plotly==5.17.0
wordcloud==1.9.2  # For generating word clouds

# Web interface
Flask-WTF==1.2.1
Werkzeug==2.3.7
Jinja2==3.1.2

# Development and testing
pytest==7.4.2
flake8==6.1.0
black==23.9.1

# Server
gunicorn==21.2.0  # For production deployment
